+++ public/model/OSMLayerFunctions.ts
@@ -1,7 +1,7 @@
-import { Map as Maplibre, LayerSpecification } from 'maplibre-gl';
+import { Map as Maplibre, LayerSpecification, SymbolLayerSpecification } from 'maplibre-gl';
 import { OSMLayerSpecification } from './mapLayerType';
-import { getMaplibreBeforeLayerId } from './layersFunctions';
 import { getLayers, hasLayer } from './map/layer_operations';
+import { getMapLanguage } from '../../common/util';
 
 interface MaplibreRef {
   current: Maplibre | null;
@@ -41,11 +41,18 @@ const updateLayerConfig = (layerConfig: OSMLayerSpecification, maplibreRef: Mapl
   handleStyleLayers(layerConfig, maplibreRef);
 };
 
-const addNewLayer = (
-  layerConfig: OSMLayerSpecification,
-  maplibreRef: MaplibreRef,
-  beforeLayerId: string | undefined
-) => {
+const setLanguage = (maplibreRef: MaplibreRef, styleLayer: LayerSpecification) => {
+  // if a layer contains label, we will set its language.
+  if (styleLayer.layout && (styleLayer as SymbolLayerSpecification).layout?.['text-field']) {
+    const language = getMapLanguage();
+    maplibreRef.current?.setLayoutProperty(styleLayer.id, 'text-field', [
+      'get',
+      'name:' + language,
+    ]);
+  }
+};
+
+const addNewLayer = (layerConfig: OSMLayerSpecification, maplibreRef: MaplibreRef) => {
   if (maplibreRef.current) {
     const { source, style } = layerConfig;
     maplibreRef.current.addSource(layerConfig.id, {
@@ -53,14 +60,14 @@ const addNewLayer = (
       url: source?.dataURL,
     });
     fetchStyleLayers(style?.styleURL).then((styleLayers: LayerSpecification[]) => {
-      const beforeMbLayerId = getMaplibreBeforeLayerId(layerConfig, maplibreRef, beforeLayerId);
       styleLayers.forEach((styleLayer) => {
         styleLayer.id = styleLayer.id + '_' + layerConfig.id;
         // TODO: Add comments on why we skip background type
         if (styleLayer.type !== 'background') {
           styleLayer.source = layerConfig.id;
         }
-        maplibreRef.current?.addLayer(styleLayer, beforeMbLayerId);
+        maplibreRef.current?.addLayer(styleLayer);
+        setLanguage(maplibreRef, styleLayer);
         maplibreRef.current?.setLayoutProperty(styleLayer.id, 'visibility', layerConfig.visibility);
         maplibreRef.current?.setLayerZoomRange(
           styleLayer.id,
@@ -83,15 +90,11 @@ const addNewLayer = (
 
 // Functions for OpenSearch maps vector tile layer
 export const OSMLayerFunctions = {
-  render: (
-    maplibreRef: MaplibreRef,
-    layerConfig: OSMLayerSpecification,
-    beforeLayerId: string | undefined
-  ) => {
+  render: (maplibreRef: MaplibreRef, layerConfig: OSMLayerSpecification) => {
     // If layer already exist in maplibre source, update layer config
     // else add new layer.
     return hasLayer(maplibreRef.current!, layerConfig.id)
       ? updateLayerConfig(layerConfig, maplibreRef)
-      : addNewLayer(layerConfig, maplibreRef, beforeLayerId);
+      : addNewLayer(layerConfig, maplibreRef);
   },
 };